transfer combinational library
