+++ src/fixed_point/combinational/tests/multiplier.py
@@ -1,9 +1,9 @@
 import pytest
 from pymtl3.stdlib.test_utils import run_test_vector_sim
 import random
-from src.fixed_point.combinational.harnesses.multiplier import MultiplierTestHarness
+from src.fixed_point.combinational.multiplier import Multiplier
 from fixedpt import Fixed
-from src.fixed_point.tools.params import mk_params, rand_fxp_spec
+from src.fixed_point.utils import mk_params, rand_fxp_spec
 
 
 # Get a random fixed point number
@@ -45,7 +45,7 @@ def test_edge(cmdline_opts, n, d, a, b):
     test_case = [("a b c* ")]
     test_case.append([a.get(), b.get(), c.get()])
 
-    run_test_vector_sim(MultiplierTestHarness(n, d, 1), test_case, cmdline_opts)
+    run_test_vector_sim(Multiplier(n, d, 1), test_case, cmdline_opts)
 
 
 @pytest.mark.parametrize(
@@ -55,10 +55,7 @@ def test_edge(cmdline_opts, n, d, a, b):
     # Extensively tests numbers with certain important bit sizes.
     sum(
         [
-            [
-                *mk_params(1, [20], n, d, slow=False),
-                *mk_params(1, [100], n, d, slow=True),
-            ]
+            mk_params(1, [100], n, d, slow=True)
             for (n, d) in [
                 (8, 4),
                 (24, 8),
@@ -74,7 +71,7 @@ def test_random(cmdline_opts, execution_number, sequence_length, n, d):
     random.seed(random.random() + execution_number)
     n, d = rand_fxp_spec(n, d)
     run_test_vector_sim(
-        MultiplierTestHarness(n, d, 1),
+        Multiplier(n, d, 1),
         gen_random_test_case(n, d, sequence_length),
         cmdline_opts,
     )
